WD-30: modular db code
